@@ -1,12 +1,13 @@
 package com.github.jacekszymanski.realcamel.entity;
 
+import jakarta.persistence.SequenceGenerator;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
 
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.MappedSuperclass;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.MappedSuperclass;
 
 @MappedSuperclass
 @Getter
@@ -16,6 +17,7 @@ public class Base {
 
   @Id
   @GeneratedValue
+  @SequenceGenerator(sequenceName = "hibernate_sequence", name = "hibernate_sequence")
   private Long id;
 
 }
